requirements.txt
decouple==0.0.7


dj-database-url==0.5.0
Django==3.1.7
django-autoslug==1.9.8
django-crispy-forms==1.11.2
django-environ==0.4.5
django-heroku==0.3.1
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.1.0
psycopg2==2.8.6
python-decouple==3.4


whitenoise==5.2.0
